feat: Experimental support for GitHub Actions v2
Run variant on pull_request/issue events and set `VARIANT_GITHUB_COMMENT` to a non-zero value, so that variant tries to send a GitHub comment to the source issue/pull request for faster feedback. The comment includes the command executed, exit status, and the stdout and full logs.

This is a succesor to https://github.com/variantdev/variant-action and complements https://github.com/variantdev/setup-variant.

`variant-action` was useless as you are unable to change the docker image for running variant.